split the rate_t, approximate rates, and modified rates into headers each is now in its own header.  This will make it easier to have modified rates be able to use StarLib rates as the underlying rate.
