+++ pynucastro/networks/tests/cxx_net_tests/_amrexastro_cxx_approx_reference/actual_rhs.H
@@ -1,5 +1,5 @@
-#ifndef actual_rhs_H
-#define actual_rhs_H
+#ifndef ACTUAL_RHS_H
+#define ACTUAL_RHS_H
 
 #include <AMReX_REAL.H>
 #include <AMReX_Array.H>
@@ -15,7 +15,10 @@
 #ifdef NEUTRINOS
 #include <neutrino.H>
 #endif
+#include <rate_type.H>
 #include <reaclib_rates.H>
+#include <approximate_rates.H>
+#include <modified_rates.H>
 #include <table_rates.H>
 #include <temperature_table_rates.H>
 #include <derived_rates.H>
@@ -137,16 +140,20 @@ void evaluate_rates(const burn_t& state,
 
     temp_tabular::fill_rates<do_T_derivatives, T>(tfactors, rate_eval);
 
-    // Calculate Derived Rates. This should go last but before approx rates.
+    // Calculate Derived Rates. This should go last but before approx
+    // rates and modified rates, since those may have ReacLibRate,
+    // TemperatureTabularRate, or DerivedRate as the underlying rate.
 
     fill_derived_rates<do_T_derivatives, T>(tfactors, rate_eval);
 
 
-    // Fill approximate rates
+    // Fill modified and approximate rates
+
+    fill_modified_rates<do_T_derivatives, T>(tfactors, rate_eval);
 
     fill_approx_rates<do_T_derivatives, T>(tfactors, state.rho, Y, rate_eval);
 
-    // Calculate tabular rates
+    // Calculate tabular weak rates
 
     [[maybe_unused]] amrex::Real rate, drate_dt, edot_nu, edot_gamma;
