+++ pynucastro/networks/tests/cxx_net_tests/_amrexastro_cxx_reference/Make.package
@@ -7,7 +7,10 @@ ifeq ($(USE_REACT),TRUE)
   CEXE_headers += interp_tools.H
   CEXE_headers += partition_functions.H
   CEXE_headers += actual_rhs.H
+  CEXE_headers += rate_type.H
   CEXE_headers += reaclib_rates.H
+  CEXE_headers += approximate_rates.H
+  CEXE_headers += modified_rates.H
   CEXE_headers += table_rates.H
   CEXE_headers += temperature_table_rates.H
   CEXE_headers += derived_rates.H
