+++ pynucastro/networks/tests/cxx_net_tests/_amrexastro_cxx_reference/temperature_table_rates.H
@@ -5,7 +5,7 @@
 
 #include <tfactors.H>
 #include <actual_network.H>
-
+#include <interp_tools.H>
 
 using namespace amrex::literals;
