+++ pynucastro/networks/tests/cxx_net_tests/_amrexastro_cxx_temptab_reference/approximate_rates.H
@@ -1,0 +1,31 @@
+#ifndef APPROXIMATE_RATES_H
+#define APPROXIMATE_RATES_H
+
+#include <AMReX.H>
+#include <AMReX_Print.H>
+
+#include <tfactors.H>
+#include <actual_network.H>
+#include <rate_type.H>
+
+using namespace Rates;
+using namespace Species;
+
+
+
+template <int do_T_derivatives, typename T>
+AMREX_GPU_HOST_DEVICE AMREX_INLINE
+void
+fill_approx_rates([[maybe_unused]] const tf_t& tfactors,
+                  [[maybe_unused]] const amrex::Real rho,
+                  [[maybe_unused]] const amrex::Array1D<amrex::Real, 1, NumSpec>& Y,
+                  [[maybe_unused]] T& rate_eval)
+{
+
+    [[maybe_unused]] amrex::Real rate{};
+    [[maybe_unused]] amrex::Real drate_dT{};
+
+
+}
+
+#endif
