+++ pynucastro/networks/tests/cxx_net_tests/_simple_cxx_reference/reaclib_rates.H
@@ -10,19 +10,19 @@ using namespace Rates;
 using namespace Species;
 
 struct rate_t {
-    Array1D<Real, 1, NumRates>  screened_rates;
+    Array1D<Real, 1, Rates::NumRates>  screened_rates;
 #ifdef SCREENING
-    Array1D<Real, 1, NumScreenPairs>  log_screen;
+    Array1D<Real, 1, Rates::NumScreenPairs>  log_screen;
 #endif
     Real enuc_weak;
 };
 
 struct rate_derivs_t {
-    Array1D<Real, 1, NumRates>  screened_rates;
-    Array1D<Real, 1, NumRates>  dscreened_rates_dT;
+    Array1D<Real, 1, Rates::NumRates>  screened_rates;
+    Array1D<Real, 1, Rates::NumRates>  dscreened_rates_dT;
 #ifdef SCREENING
-    Array1D<Real, 1, NumScreenPairs>  log_screen;
-    Array1D<Real, 1, NumScreenPairs>  dlog_screen_dT;
+    Array1D<Real, 1, Rates::NumScreenPairs>  log_screen;
+    Array1D<Real, 1, Rates::NumScreenPairs>  dlog_screen_dT;
 #endif
     Real enuc_weak;
 };
