+++ pynucastro/templates/amrexastro-cxx-microphysics/temperature_table_rates.H.template
@@ -5,7 +5,7 @@
 
 #include <tfactors.H>
 #include <actual_network.H>
-
+#include <interp_tools.H>
 
 using namespace amrex::literals;
